update manager -> scenario namespace rename in files
