+++ cmake/headers.cmake
@@ -26,25 +26,25 @@ include/warthog/heuristic/manhattan_heuristic.h
 include/warthog/heuristic/octile_heuristic.h
 include/warthog/heuristic/zero_heuristic.h
 
+include/warthog/io/bittable_serialize.h
 include/warthog/io/fwd.h
-include/warthog/io/grid.h
 include/warthog/io/grid_trace.h
 include/warthog/io/log.h
 include/warthog/io/observer.h
 include/warthog/io/posthoc_trace.h
-include/warthog/io/scenario.h
+include/warthog/io/scenario_serialize.h
 include/warthog/io/serialize_base.h
 include/warthog/io/stream_observer.h
-
-include/warthog/manager/experiment.h
-include/warthog/manager/grid_patch_set.h
-include/warthog/manager/scenario_manager.h
-include/warthog/manager/scenario_runner.h
 
 include/warthog/memory/arraylist.h
 include/warthog/memory/bittable.h
 include/warthog/memory/cpool.h
 include/warthog/memory/node_pool.h
+
+include/warthog/scenario/experiment.h
+include/warthog/scenario/grid_patch_set.h
+include/warthog/scenario/scenario_manager.h
+include/warthog/scenario/scenario_runner.h
 
 include/warthog/search/dummy_filter.h
 include/warthog/search/dummy_listener.h
